+++ apis/member-api-server/src/main/java/com/moviemang/member/service/MemberService.java
@@ -1,13 +1,13 @@
 package com.moviemang.member.service;
 
 import com.moviemang.coreutils.common.response.CommonResponse;
-import com.moviemang.datastore.domain.MailCertificationDto;
-import com.moviemang.datastore.entity.maria.Member;
+import com.moviemang.datastore.dto.MailCertificationDto;
+import com.moviemang.datastore.dto.member.MemberJoinDto;
 
 @SuppressWarnings("rawtypes")
 public interface MemberService {
 	CommonResponse checkMailCertification(MailCertificationDto certificationDto);
-    public Member regist(Member member);
+    public CommonResponse regist(MemberJoinDto member);
     public CommonResponse checkEmail(String email);
     public CommonResponse checkNick(String nick);
 }
